+++ BE/app.js
@@ -12,12 +12,10 @@ var projectsRouter = require("./routes/projects/projects");
 var travelRouter = require("./routes/travel/travel");
 var commonRouter = require("./routes/common/common");
 var voiceRouter = require("./routes/voicetalk/voicetalk");
-var sharememoRouter = require("./routes/shareMemo/shareMemo");
 var mongodb = require("dotenv").config();
 var fs = require("fs");
 
 voiceRouter;
-sharememoRouter;
 
 var app = express();
 // [원영] 소켓 서버 추가
@@ -107,50 +105,50 @@ app.use("/common", commonRouter);
 
 // [수연] share-memo with collaborative cursors
 // create and start server on 7899 port by default
-var OkdbServer = require("okdb-server");
-var options = {
-  cors: {
-    enabled: true,
-    allowedOrigins: `https://${process.env.REACT_APP_SERVER_IP}:3000`,
-  },
-};
-var okdb = new OkdbServer(options);
+// var OkdbServer = require("okdb-server");
+// var options = {
+//   cors: {
+//     enabled: true,
+//     allowedOrigins: `https://${process.env.REACT_APP_SERVER_IP}:3000`,
+//   },
+// };
+// var okdb = new OkdbServer(options);
 
-// sample authentication, e.g. should validate your own auth token
-let nameIdx = 0;
-try {
-  okdb.handlers().auth(({ myNickname, selectedIndex }) => {
-    if (myNickname) {
-      console.log("auth attempt for ", myNickname, " success");
-      const userName = myNickname;
-      const userId = "1" + nameIdx;
-      nameIdx = (nameIdx + 1) % 10;
-      return { id: userId, name: userName, selectedIndex: selectedIndex };
-    }
-    console.log("auth attempt for ", myNickname, " failed");
-    return null;
-  });
-} catch (err) {
-  console.log(err);
-}
+// // sample authentication, e.g. should validate your own auth token
+// let nameIdx = 0;
+// try {
+//   okdb.handlers().auth(({ myNickname, selectedIndex }) => {
+//     if (myNickname) {
+//       console.log("auth attempt for ", myNickname, " success");
+//       const userName = myNickname;
+//       const userId = "1" + nameIdx;
+//       nameIdx = (nameIdx + 1) % 10;
+//       return { id: userId, name: userName, selectedIndex: selectedIndex };
+//     }
+//     console.log("auth attempt for ", myNickname, " failed");
+//     return null;
+//   });
+// } catch (err) {
+//   console.log(err);
+// }
 
-// Handling Ctrl-C (workaround for Windows)
-if (process.platform === "win32") {
-  var rl = require("readline").createInterface({
-    input: process.stdin,
-    output: process.stdout,
-  });
+// // Handling Ctrl-C (workaround for Windows)
+// if (process.platform === "win32") {
+//   var rl = require("readline").createInterface({
+//     input: process.stdin,
+//     output: process.stdout,
+//   });
 
-  rl.on("SIGINT", function () {
-    process.emit("SIGINT");
-  });
-}
-//graceful shutdown on Ctrl-C (all other platforms)
-process.on("SIGINT", function () {
-  okdb.stop(() => {
-    console.log("server stopped");
-    process.exit();
-  });
-});
+//   rl.on("SIGINT", function () {
+//     process.emit("SIGINT");
+//   });
+// }
+// //graceful shutdown on Ctrl-C (all other platforms)
+// process.on("SIGINT", function () {
+//   okdb.stop(() => {
+//     console.log("server stopped");
+//     process.exit();
+//   });
+// });
 
 module.exports = app;
